Alow building against older FFmpeg versions.
Although, building against a new(er) FFmpeg version would be preferable, in
order to match with the upstream static builds, there seems to be no recent
version around for CentOS 7 which doesn't link against non-free or possibly
patented libraries.

As of this, it was decided to build against the FFmpeg version provided by the
Nux Dextop repository. Despite the legacy FFmpeg version contained in this
repository, Nux Dextop is widely available and we already have a working FFmpeg
build dependency on OBS.

In case someone would like to build the spec file against a newer FFmpeg
version, the [negativo17 - Multimedia
(epel-multimedia)](https://negativo17.org/multimedia/) repository can be used.